+++ R/plus.R
@@ -12,9 +12,13 @@ setMethod(
   ),
   function(e1, e2) {
     e2_token <- e2@object$tokens
-    e2_name <- e2@object$names
+    e2_name <- e2@object$name
     e2_opacity <- e2@object$opacity
     e2_shown <- e2@object$shown
+    e2_min <- e2@object$min
+    e2_max <- e2@object$max
+    e2_pal <- e2@object$palette
+    e2_legend <- e2@object$legend
 
     for (x in seq_len(length(e2_name))) {
       e1@map <- e1@map %>%
@@ -26,7 +30,19 @@ setMethod(
         ee_mapViewLayersControl(names = e2_name[x]) %>%
         leaflet::hideGroup(if (!e2_shown[x]) e2_name[x] else NULL)
     }
-    return(e1)
+
+    if (isTRUE(e2_legend)) {
+      e1@map <- e1@map %>%
+        leaflet::addLegend(
+          position = "bottomright",
+          pal = e2_pal,
+          values = c(e2_min, e2_max),
+          opacity = 1,
+          title = e2_name
+        )
+    } else {
+      e1
+    }
   }
 )